Update vue component tests.
The npm package name for vue test utils changed, so all tests' imports have been.

At some point, the naming of the components and component directories changed, so the tests have been updated to import the proper components.

These tests were also no longer testing existing functionality, so all have been marked as skipped. The tests have been left in
place as a marker of what will need to be tested, and may be updated or deleted and replaced with working tests as components are updated.